SignLanguageProject/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
tensorflow>=2.11.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyttsx3>=2.90
absl-py>=1.0.0
